+++ CHANGELOG.md
@@ -16,4 +16,6 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
  - Enabled peer scoring by default. Can be disabled explicitly using `--Xp2p-gossip-scoring-enabled=false`
+ - When failovers are configured, a validator client will perform a readiness check on startup to avoid retrieving validator statuses from a node which is not ready
+
 ### Bug Fixes